Use `cuda-python` version to select CUDA version
Instead of having the CUDA version defined to build against in various
places, consolidate down to just the `version` used by `cuda-python` at
the beginning of the recipe. The propagate this value into how the
`skip`s are configured to select the relevant CUDA version. Also
leverage this in how `cuda-version` is constrained in the `requirements`
during the build phase. This should simplify the upgrade process of
`cuda-python` for new versions (including major CUDA version bumps).